- move view models to model package - fix drawer not showing switch account/sign out
