+++ app/src/main/java/com/jerboa/ui/components/comment/reply/CommentReplyActivity.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.jerboa.api.ApiState
-import com.jerboa.db.AccountViewModel
+import com.jerboa.model.AccountViewModel
 import com.jerboa.model.CommentReplyViewModel
 import com.jerboa.model.ReplyItem
 import com.jerboa.model.SiteViewModel
