+++ app/src/main/java/com/jerboa/ui/components/post/edit/PostEditActivity.kt
@@ -21,9 +21,9 @@ import com.jerboa.api.ApiState
 import com.jerboa.api.uploadPictrsImage
 import com.jerboa.datatypes.types.EditPost
 import com.jerboa.datatypes.types.PostView
-import com.jerboa.db.AccountViewModel
 import com.jerboa.db.entity.Account
 import com.jerboa.imageInputStreamFromUri
+import com.jerboa.model.AccountViewModel
 import com.jerboa.model.PostEditViewModel
 import com.jerboa.ui.components.common.InitializeRoute
 import com.jerboa.ui.components.common.LoadingBar
