+++ app/src/main/java/com/jerboa/ui/components/privatemessage/PrivateMessageReplyActivity.kt
@@ -18,7 +18,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.jerboa.api.ApiState
 import com.jerboa.datatypes.types.PrivateMessageView
-import com.jerboa.db.AccountViewModel
+import com.jerboa.model.AccountViewModel
 import com.jerboa.model.PrivateMessageReplyViewModel
 import com.jerboa.model.SiteViewModel
 import com.jerboa.ui.components.common.InitializeRoute
