+++ app/src/main/java/com/jerboa/ui/components/settings/account/AccountSettingsActivity.kt
@@ -11,8 +11,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.res.stringResource
 import androidx.navigation.NavController
 import com.jerboa.R
-import com.jerboa.db.AccountViewModel
 import com.jerboa.model.AccountSettingsViewModel
+import com.jerboa.model.AccountViewModel
 import com.jerboa.model.SiteViewModel
 import com.jerboa.ui.components.common.SimpleTopAppBar
 import com.jerboa.ui.components.common.getCurrentAccount
